+++ tensornetwork/block_sparse/blocksparsetensor_test.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
-# pylint: disable=line-too-long
-from tensornetwork.block_sparse.charge import U1Charge, fuse_charges, charge_equal, fuse_ndarrays, fuse_ndarray_charges, BaseCharge, Z2Charge
-from tensornetwork.block_sparse.index import Index
-from tensornetwork.block_sparse.blocksparsetensor import ChargeArray, BlockSparseTensor
+
+from tensornetwork.block_sparse.charge_new import (U1Charge, fuse_charges, charge_equal, fuse_ndarrays, fuse_ndarray_charges, BaseCharge, Z2Charge)
+from tensornetwork.block_sparse.index_new import Index
+from tensornetwork.block_sparse.blocksparsetensor_new import (ChargeArray, BlockSparseTensor)
 
 np_dtypes = [np.float64, np.complex128]
 np_tensordot_dtypes = [np.float64, np.complex128]
@@ -12,21 +12,21 @@ np_tensordot_dtypes = [np.float64, np.complex128]
 def get_charge(chargetype, num_charges, D):
   if chargetype == "U1":
     return BaseCharge(
-        np.random.randint(-5, 6, (num_charges, D)),
-        charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6,  D).astype(np.int16)for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   if chargetype == "Z2":
     return BaseCharge(
-        np.random.randint(0, 2, (num_charges, D)),
-        charge_types=[Z2Charge] * num_charges)
+      [np.random.randint(0, 2, D).astype(np.int8)for _ in range(num_charges)],
+      charge_types=[[Z2Charge]] * num_charges)
   if chargetype == "mixed":
     n1 = num_charges // 2 if num_charges > 1 else 1
     c = BaseCharge(
-        np.random.randint(-5, 6, (n1, D)), charge_types=[U1Charge] * n1)
+      [np.random.randint(-5, 6,  D).astype(np.int16)for _ in range(n1)], charge_types=[[U1Charge]] * n1)
 
     if num_charges > 1:
       n2 = num_charges - n1
       c = c @ BaseCharge(
-          np.random.randint(0, 2, (n2, D)), charge_types=[Z2Charge] * n2)
+          [np.random.randint(0, 2,  D).astype(np.int8)for _ in range(n2)], charge_types=[[Z2Charge]] * n2)
 
     return c
   return None
@@ -267,7 +267,7 @@ def test_BlockSparseTensor_init():
   ]
   fused = fuse_charges(charges, flows)
   data = np.random.uniform(
-      0, 1, size=len(np.nonzero(fused == np.zeros((1, 1)))[0]))
+    0, 1, size=len(np.nonzero(fused == fused.identity_charges(1))[0]))
   order = [[n] for n in range(rank)]
   arr = BlockSparseTensor(data, charges, flows, order=order)
   np.testing.assert_allclose(data, arr.data)
@@ -276,7 +276,7 @@ def test_BlockSparseTensor_init():
   for c1, c2 in zip(charges, arr._charges):
     assert charge_equal(c1, c2)
   data = np.random.uniform(
-      0, 1, size=len(np.nonzero(fused == np.zeros((1, 1)))[0]) + 1)
+      0, 1, size=len(np.nonzero(fused == fused.identity_charges(1))[0]) + 1)
   with pytest.raises(ValueError):
     arr = BlockSparseTensor(
         data, charges, flows, order=order, check_consistency=True)
@@ -389,7 +389,7 @@ def test_todense(num_charges, chargetype):
   flows = np.random.choice([True, False], size=rank, replace=True)
   charges = [get_charge(chargetype, num_charges, Ds[n]) for n in range(rank)]
   fused = fuse_charges(charges, flows)
-  mask = fused == np.zeros((num_charges, 1))
+  mask = fused == fused.identity_charges(1)
   inds = np.nonzero(mask)[0]
   inds2 = np.nonzero(np.logical_not(mask))[0]
   indices = [Index(charges[n], flows[n]) for n in range(rank)]
@@ -408,7 +408,7 @@ def test_fromdense(num_charges, chargetype):
   flows = np.random.choice([True, False], size=rank, replace=True)
   charges = [get_charge(chargetype, num_charges, Ds[n]) for n in range(rank)]
   fused = fuse_charges(charges, flows)
-  mask = fused == np.zeros((num_charges, 1))
+  mask = fused == fused.identity_charges(1)
   inds = np.nonzero(mask)[0]
   inds2 = np.nonzero(np.logical_not(mask))[0]
   indices = [Index(charges[n], flows[n]) for n in range(rank)]
@@ -633,15 +633,15 @@ def test_matmul_raises():
   is1 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 5, (num_charges, Ds1[n]), dtype=np.int16),
-              charge_types=[U1Charge] * num_charges), False) for n in range(3)
+              [np.random.randint(-5, 5, Ds1[n], dtype=np.int16) for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), False) for n in range(3)
   ]
   is2 = [
       is1[1].copy().flip_flow(),
       Index(
           BaseCharge(
-              np.random.randint(-5, 5, (num_charges, 150), dtype=np.int16),
-              charge_types=[U1Charge] * num_charges), False)
+              [np.random.randint(-5, 5, 150, dtype=np.int16) for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), False)
   ]
   tensor1 = BlockSparseTensor.random(is1, dtype=dtype)
   tensor2 = BlockSparseTensor.random(is2, dtype=dtype)
