+++ tensornetwork/block_sparse/charge_test.py
@@ -1,36 +1,261 @@
 import numpy as np
+from tensornetwork.block_sparse.charge_new import (collapse, expand, BaseCharge,
+                                                   U1Charge, Z2Charge, ZNCharge,
+                                                   intersect, fuse_ndarrays,
+                                                   fuse_degeneracies,
+                                                   fuse_charges)
 import pytest
-# pylint: disable=line-too-long
-from tensornetwork.block_sparse.charge import BaseCharge, intersect, fuse_ndarrays, U1Charge, fuse_degeneracies, fuse_charges, Z2Charge, ZNCharge
+
+# pylint: disable=unbalanced-tuple-unpacking
+def test_charge_collapse_expand_1():
+  np.random.seed(10)
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  [a1f, b1f] = expand(a1b1, [[a1.dtype], [b1.dtype]]) 
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  assert a1b1.dtype == np.int32
+  a1f, b1f = expand(a1b1, [[a1.dtype], [b1.dtype]])
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  assert a1b1.dtype == np.int32
+  a1f, b1f = expand(a1b1, [[a1.dtype], [b1.dtype]])
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  c1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  a1b1c1 = collapse([a1, b1, c1], [[a1.dtype], [b1.dtype], [c1.dtype]])
+  assert a1b1c1.dtype == np.int32
+  a1f, b1f, c1f = expand(a1b1c1, [[a1.dtype], [b1.dtype], [c1.dtype]])
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+  np.testing.assert_allclose(c1f, c1)
+
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  c1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  a1b1c1 = collapse([a1, b1, c1], [[a1.dtype], [b1.dtype], [c1.dtype]])
+  assert a1b1c1.dtype == np.int32
+  a1f, b1f, c1f = expand(a1b1c1, [[a1.dtype], [b1.dtype], [c1.dtype]])
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+  np.testing.assert_allclose(c1f, c1)
+
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  c1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  a1b1c1 = collapse([a1, b1, c1], [[a1.dtype], [b1.dtype], [c1.dtype]])
+  assert a1b1c1.dtype == np.int32
+  a1f, b1f, c1f = expand(a1b1c1, [[a1.dtype], [b1.dtype], [c1.dtype]])
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+  np.testing.assert_allclose(c1f, c1)
+
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  c1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  a1b1c1 = collapse([a1, b1, c1], [[a1.dtype], [b1.dtype], [c1.dtype]])
+  assert a1b1c1.dtype == np.int32
+  a1f, b1f, c1f = expand(a1b1c1, [[a1.dtype], [b1.dtype], [c1.dtype]])
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(b1f, b1)
+  np.testing.assert_allclose(c1f, c1)
+
+
+def test_charge_collapse_expand_2():
+  np.random.seed(10)
+  a = np.random.randint(-10, 10, 10000).astype(np.int8)
+  b = np.random.randint(-10, 10, 10000).astype(np.int16)
+  c = np.random.randint(-10, 10, 10000).astype(np.int8)
+
+  r1 = np.left_shift(a.astype(np.int32), 24) + np.left_shift(
+      b.astype(np.int32), 8) + c.astype(np.int32)
+  r2 = np.left_shift(a.astype(np.int32), 24) + (
+      np.left_shift(b.astype(np.int32), 8) + c.astype(np.int32))
+  np.testing.assert_allclose(r1, r2)
+
+  cf = np.bitwise_and(r1, 2**8 - 1).astype(np.int8)
+  np.testing.assert_allclose(cf, c)
+  r1 = np.right_shift(r1 - cf.astype(np.int32), 8)
+  bf = np.bitwise_and(r1, 2**16 - 1).astype(np.int16)
+  np.testing.assert_allclose(bf, b)
+  af = np.right_shift(r1 - bf.astype(np.int32), 16).astype(np.int8)
+  np.testing.assert_allclose(af, a)
+
+  bc = collapse([b, c], [[b.dtype], [c.dtype]])
+  abc = collapse([a, bc], [[a.dtype], [b.dtype, c.dtype]])
+  abc2 = collapse([a, b, c], [[a.dtype], [b.dtype], [c.dtype]])
+  np.testing.assert_allclose(abc, r2)
+  np.testing.assert_allclose(abc2, r2)
+
+
+def test_charge_collapse_expand_3():
+  np.random.seed(10)
+  a1 = np.random.randint(-10, 10, 10000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 10000).astype(np.int8)
+  a2 = np.random.randint(-10, 10, 10000).astype(np.int8)
+  b2 = np.random.randint(-10, 10, 10000).astype(np.int8)
+
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  a2b2 = collapse([a2, b2], [[a2.dtype], [b2.dtype]])
+
+  a1b1a2b2 = collapse([a1b1, a2b2],
+                      [[a1.dtype, b1.dtype], [a2.dtype, b2.dtype]])
+  a1b1_, a2b2_ = expand(a1b1a2b2, [[a1.dtype, b1.dtype], [a2.dtype, b2.dtype]])
+  a1f, b1f = expand(a1b1_, [[a1.dtype], [b1.dtype]])
+  a2f, b2f = expand(a2b2_, [[a2.dtype], [b2.dtype]])
+
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(a2f, a2)
+  np.testing.assert_allclose(b1f, b1)
+  np.testing.assert_allclose(b2f, b2)
+
+
+def test_charge_collapse_expand_4():
+  np.random.seed(10)
+  a1 = np.random.randint(-10, 10, 10000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 10000).astype(np.int16)
+  a2 = np.random.randint(-10, 10, 10000).astype(np.int8)
+
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  a1b1a2 = collapse([a1b1, a2], [[a1.dtype, b1.dtype], [a2.dtype]])
+  a1b1_, a2f = expand(a1b1a2, [[a1.dtype, b1.dtype], [a2.dtype]])
+  a1f, b1f = expand(a1b1_, [[a1.dtype], [b1.dtype]])
+
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(a2f, a2)
+  np.testing.assert_allclose(b1f, b1)
+
+
+def test_charge_collapse_expand_5():
+  np.random.seed(10)
+  a = np.random.randint(-10, 10, 10000).astype(np.int8)
+  b = np.random.randint(-10, 10, 10000).astype(np.int16)
+  c = np.random.randint(-10, 10, 10000).astype(np.int8)
+
+  bc = collapse([b, c], [[b.dtype], [c.dtype]])
+  abc = collapse([a, bc], [[a.dtype], [b.dtype, c.dtype]])
+  abc2 = collapse([a, b, c], [[a.dtype], [b.dtype], [c.dtype]])
+  r1 = np.left_shift(a.astype(np.int32), 24) + np.left_shift(
+      b.astype(np.int32), 8) + c.astype(np.int32)
+  np.testing.assert_allclose(abc, r1)
+  np.testing.assert_allclose(abc2, abc)
+
+  af, bf, cf = expand(abc, [[a.dtype], [b.dtype], [c.dtype]])
+  np.testing.assert_allclose(af, a)
+  np.testing.assert_allclose(bf, b)
+  np.testing.assert_allclose(cf, c)
+  af, bcf = expand(abc, [[a.dtype], [b.dtype, c.dtype]])
+  bf, cf = expand(bcf, [[b.dtype], [c.dtype]])
+  np.testing.assert_allclose(af, a)
+  np.testing.assert_allclose(bf, b)
+  np.testing.assert_allclose(cf, c)
+
+
+def test_charge_collapse_expand_6():
+  np.random.seed(10)
+  a1 = np.random.randint(-10, 10, 1000000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 1000000).astype(np.int16)
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  a2 = np.random.randint(-10, 10, 1000000).astype(np.int8)
+  b2 = np.random.randint(-10, 10, 1000000).astype(np.int16)
+  a2b2 = collapse([a2, b2], [[a2.dtype], [b2.dtype]])
+  a1b1a2b2 = collapse([a1b1, a2b2],
+                      [[a1.dtype, b1.dtype], [a2.dtype, b2.dtype]])
+  a1b1_, a2b2_ = expand(a1b1a2b2, [[a1.dtype, b1.dtype], [a2.dtype, b2.dtype]])
+  a1f, b1f = expand(a1b1_, [[a1.dtype], [b1.dtype]])
+  a2f, b2f = expand(a2b2_, [[a2.dtype], [b2.dtype]])
+
+  np.testing.assert_allclose(a1f, a1)
+  np.testing.assert_allclose(a2f, a2)
+  np.testing.assert_allclose(b1f, b1)
+  np.testing.assert_allclose(b2f, b2)
+
+
+def test_charge_collapse_expand_7():
+  np.random.seed(10)
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  a2 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b2 = np.random.randint(-10, 10, 100000).astype(np.int16)
+
+  a1b1 = collapse([a1, b1], [[a1.dtype], [b1.dtype]])
+  a2b2 = collapse([a2, b2], [[a2.dtype], [b2.dtype]])
+  res = a1b1 + a2b2
+  resa, resb = expand(res, [[a1.dtype], [b1.dtype]])
+  np.testing.assert_allclose(resa, a1 + a2)
+  np.testing.assert_allclose(resb, b1 + b2)
+
+
+def test_charge_collapse_expand_8():
+  np.random.seed(10)
+  a1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b1 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  c1 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  a2 = np.random.randint(-10, 10, 100000).astype(np.int8)
+  b2 = np.random.randint(-10, 10, 100000).astype(np.int16)
+  c2 = np.random.randint(-10, 10, 100000).astype(np.int8)
+
+  a1b1c1 = collapse([a1, b1, c1], [[a1.dtype], [b1.dtype], [c1.dtype]])
+  a2b2c2 = collapse([a2, b2, c2], [[a2.dtype], [b2.dtype], [c2.dtype]])
+  res = a1b1c1 + a2b2c2
+  resa, _ = expand(res, [[a1.dtype], [b1.dtype, c1.dtype]])
+  resb, resc = expand(res, [[b1.dtype], [c1.dtype]])
+  np.testing.assert_allclose(resa, a1 + a2)
+  np.testing.assert_allclose(resb, b1 + b2)
+  np.testing.assert_allclose(resc, c1 + c2)
+  resa2, resb2, resc2 = expand(res, [[a1.dtype], [b1.dtype], [c1.dtype]])
+  np.testing.assert_allclose(resa2, a1 + a2)
+  np.testing.assert_allclose(resb2, b1 + b2)
+  np.testing.assert_allclose(resc2, c1 + c2)
 
 
 def test_BaseCharge_charges():
   D = 100
   B = 6
   np.random.seed(10)
-  charges = np.random.randint(-B // 2, B // 2 + 1, (2, D)).astype(np.int16)
-
+  charges = [
+      np.random.randint(-B // 2, B // 2 + 1, D).astype(np.int16)
+      for _ in range(2)
+  ]
   q1 = BaseCharge(charges)
-  np.testing.assert_allclose(q1.charges, charges)
+  q1.expand_charge_types()
+  np.testing.assert_allclose(
+      np.stack(q1.charges, axis=0), np.stack(charges, axis=0))
 
 
 def test_BaseCharge_generic():
   D = 300
   B = 5
   np.random.seed(10)
-  q = np.random.randint(-B // 2, B // 2 + 1, (2, D)).astype(np.int16)
-  unique = np.unique(q, axis=1)
-  Q = BaseCharge(charges=q)
+  charges = [
+      np.random.randint(-B // 2, B // 2 + 1, D).astype(np.int16)
+      for _ in range(2)
+  ]
+  Q = BaseCharge(charges)
   assert Q.dim == 300
   assert Q.num_symmetries == 2
-  assert Q.num_unique == unique.shape[1]
 
 
 def test_BaseCharge_len():
   D = 300
   B = 5
   np.random.seed(10)
-  q = np.random.randint(-B // 2, B // 2 + 1, (2, D)).astype(np.int16)
+  q = [
+      np.random.randint(-B // 2, B // 2 + 1, D).astype(np.int16)
+      for _ in range(2)
+  ]
   Q = BaseCharge(charges=q)
   assert len(Q) == 300
 
@@ -39,38 +264,38 @@ def test_BaseCharge_copy():
   D = 300
   B = 5
   np.random.seed(10)
-  q = np.random.randint(-B // 2, B // 2 + 1, (2, D)).astype(np.int16)
+  q = [
+      np.random.randint(-B // 2, B // 2 + 1, D).astype(np.int16)
+      for _ in range(2)
+  ]
   Q = BaseCharge(charges=q)
   Qcopy = Q.copy()
-  assert Q.charge_labels is not Qcopy.charge_labels
-  assert Q.unique_charges is not Qcopy.unique_charges
-  np.testing.assert_allclose(Q.charge_labels, Qcopy.charge_labels)
-  np.testing.assert_allclose(Q.unique_charges, Qcopy.unique_charges)
+  for n, c in enumerate(Q.charges):
+    assert c is not Qcopy.charges[n]
+  np.testing.assert_allclose(np.stack(Q.charges), Qcopy.charges)
 
 
 def test_BaseCharge_unique():
   D = 3000
   B = 5
   np.random.seed(10)
-  q = np.random.randint(-B // 2, B // 2 + 1, (2, D)).astype(np.int16)
-  Q = BaseCharge(charges=q, charge_types=[U1Charge, U1Charge])
+  q = [
+      np.random.randint(-B // 2, B // 2 + 1, D).astype(np.int16)
+      for _ in range(2)
+  ]
+  Q = BaseCharge(charges=q, charge_types=[[U1Charge], [U1Charge]])
   expected = np.unique(
-      q, return_index=True, return_inverse=True, return_counts=True, axis=1)
+      np.stack(q, axis=1),
+      return_index=True,
+      return_inverse=True,
+      return_counts=True,
+      axis=0)
   actual = Q.unique(return_index=True, return_inverse=True, return_counts=True)
-  assert np.all(actual[0].charges == expected[0])
+  actual[0].expand_charge_types()
+  np.testing.assert_allclose(np.stack(actual[0].charges, axis=1), expected[0])
   assert np.all(actual[1] == expected[1])
   assert np.all(actual[2] == expected[2])
   assert np.all(actual[3] == expected[3])
-
-
-def test_BaseCharge_unique_sort():
-  np.random.seed(10)
-  unique = np.array([1, 0, -1])
-  labels = np.random.randint(0, 3, 100)
-  Q = U1Charge(charges=unique, charge_labels=labels)
-  actual = Q.unique(
-      return_index=True, return_inverse=True, return_counts=True, sort=False)
-  np.testing.assert_allclose(actual[0].unique_charges, [[1, 0, -1]])
 
 
 def test_intersect_1():
@@ -145,7 +370,7 @@ def test_Charge_charges(chargetype, B0, B1):
   np.random.seed(10)
   charges = np.random.randint(B0, B1 + 1, D).astype(np.int16)
   q1 = chargetype(charges)
-  assert np.all(q1.charges == charges)
+  np.testing.assert_allclose(q1.charges[0], charges)
 
 
 @pytest.mark.parametrize('chargetype, B0, B1,sign', [(U1Charge, -5, 5, -1),
@@ -156,7 +381,7 @@ def test_Charge_dual(chargetype, B0, B1, sign):
   charges = np.random.randint(B0, B1 + 1, D).astype(np.int16)
 
   q1 = chargetype(charges)
-  assert np.all(q1.dual(True).charges == sign * charges)
+  assert np.all(q1.dual(True).charges[0] == sign * charges)
 
 
 @pytest.mark.parametrize('n', list(range(2, 12)))
@@ -166,7 +391,7 @@ def test_Charge_dual_zncharges(n):
   np.random.seed(10)
   charges = np.random.randint(0, n, D).astype(np.int16)
   q1 = chargetype(charges)
-  assert np.all(q1.dual(True).charges == (n - charges) % n)
+  assert np.all(q1.dual(True).charges[0] == (n - charges) % n)
 
 
 def test_Z2Charge_raises():
@@ -176,37 +401,30 @@ def test_Z2Charge_raises():
     Z2Charge(charges)
 
 
-def get_charges(B0, B1, D, num_charges):
+def get_charges(B0, B1, D, num_legs):
   return [
-      np.random.randint(B0, B1 + 1, D).astype(np.int16)
-      for _ in range(num_charges)
+      np.random.randint(B0, B1 + 1, D).astype(np.int16) for _ in range(num_legs)
   ]
 
 
-def fuse_many_charges(num_charges,
-                      num_charge_types,
-                      seed,
-                      D,
-                      B,
-                      use_flows=False):
+def fuse_many_charges(num_legs, num_charge_types, seed, D, B, use_flows=False):
   np.random.seed(seed)
   if use_flows:
-    flows = np.random.choice([True, False], num_charges, replace=True)
+    flows = np.random.choice([True, False], num_legs, replace=True)
   else:
-    flows = np.asarray([False] * num_charges)
-  np_flows = np.ones(num_charges, dtype=np.int16)
+    flows = np.asarray([False] * num_legs)
+  np_flows = np.ones(num_legs, dtype=np.int16)
   np_flows[flows] = -1
   charges = [
-      get_charges(-B // 2, B // 2, D, num_charge_types)
-      for _ in range(num_charges)
+      get_charges(-B // 2, B // 2, D, num_charge_types) for _ in range(num_legs)
   ]
   fused = [
       fuse_ndarrays([charges[n][m] * np_flows[n]
-                     for n in range(num_charges)])
+                     for n in range(num_legs)])
       for m in range(num_charge_types)
   ]
-  final_charges = [U1Charge(charges[n][0]) for n in range(num_charges)]
-  for n in range(num_charges):
+  final_charges = [U1Charge(charges[n][0]) for n in range(num_legs)]
+  for n in range(num_legs):
     for m in range(1, num_charge_types):
       final_charges[n] = final_charges[n] @ U1Charge(charges[n][m])
   np_target_charges = np.random.randint(-B, B, num_charge_types, dtype=np.int16)
@@ -218,11 +436,11 @@ def fuse_many_charges(num_charges,
   for m in range(1, num_charge_types):
     target = target @ target_charges[m]
   final = final_charges[0] * flows[0]
-  for n in range(1, num_charges):
+  for n in range(1, num_legs):
     final = final + final_charges[n] * flows[n]
 
   nz_1 = np.nonzero(final == target)[0]
-  masks = [fused[m] == target.charges[m, 0] for m in range(num_charge_types)]
+  masks = [fused[m] == np_target_charges[m] for m in range(num_charge_types)]
 
   #pylint: disable=no-member
   nz_2 = np.nonzero(np.logical_and.reduce(masks))[0]
@@ -230,12 +448,14 @@ def fuse_many_charges(num_charges,
 
 
 @pytest.mark.parametrize('use_flows', [True, False])
-@pytest.mark.parametrize('num_charges, num_charge_types, D, B',
-                         [(2, 1, 1000, 6), (2, 2, 1000, 6), (3, 1, 100, 6),
-                          (3, 2, 100, 6), (3, 3, 100, 6)])
-def test_U1Charge_fusion(num_charges, num_charge_types, D, B, use_flows):
+@pytest.mark.parametrize('num_legs, num_charge_types, D, B', [(2, 1, 1000, 6),
+                                                              (2, 2, 1000, 6),
+                                                              (3, 1, 100, 6),
+                                                              (3, 2, 100, 6),
+                                                              (3, 3, 100, 6)])
+def test_U1Charge_fusion(num_legs, num_charge_types, D, B, use_flows):
   nz_1, nz_2 = fuse_many_charges(
-      num_charges=num_charges,
+      num_legs=num_legs,
       num_charge_types=num_charge_types,
       seed=20,
       D=D,
@@ -247,29 +467,39 @@ def test_U1Charge_fusion(num_charges, num_charge_types, D, B, use_flows):
 
 
 def test_BaseCharge_intersect():
-  q1 = np.array([[0, 1, 2, 0, 6], [2, 3, 4, -1, 4]])
-  q2 = np.array([[0, -2, 6], [2, 3, 4]])
+  q1 = [
+      np.array([0, 1, 2, 0, 6], dtype=np.int16),
+      np.array([2, 3, 4, -1, 4], dtype=np.int16)
+  ]
+  q2 = [
+      np.array([0, -2, 6], dtype=np.int16),
+      np.array([2, 3, 4], dtype=np.int16)
+  ]
   Q1 = BaseCharge(charges=q1)
   Q2 = BaseCharge(charges=q2)
   res = Q1.intersect(Q2)
-  np.testing.assert_allclose(res.charges, np.asarray([[0, 6], [2, 4]]))
-
-
-def test_BaseCharge_intersect_2():
-  c1 = U1Charge(np.array([1, 0, -1]), charge_labels=np.array([2, 0, 1]))
-  c2 = U1Charge(np.array([-1, 0, 1]))
-  res = c1.intersect(c2)
-  np.testing.assert_allclose(res.charges, [[-1, 0, 1]])
+  res.expand_charge_types()
+  np.testing.assert_allclose(
+      np.stack(res.charges, axis=0), np.asarray([[0, 6], [2, 4]]))
 
 
 def test_BaseCharge_intersect_return_indices():
-  q1 = np.array([[0, 1, 2, 0, 6], [2, 3, 4, -1, 4]])
-  q2 = np.array([[-2, 0, 6], [3, 2, 4]])
+  q1 = [
+      np.array([0, 1, 2, 0, 6], dtype=np.int16),
+      np.array([2, 3, 4, -1, 4], dtype=np.int16)
+  ]
+  q2 = [
+      np.array([-2, 0, 6], dtype=np.int16),
+      np.array([3, 2, 4], dtype=np.int16)
+  ]
+
   Q1 = BaseCharge(charges=q1)
   Q2 = BaseCharge(charges=q2)
   res, i1, i2 = Q1.intersect(Q2, return_indices=True)
   #res, i1, i2 = intersect(q1, q2, axis=1, return_indices=True)
-  np.testing.assert_allclose(res.charges, np.asarray([[0, 6], [2, 4]]))
+  res.expand_charge_types()
+  np.testing.assert_allclose(
+      np.stack(res.charges, axis=0), np.asarray([[0, 6], [2, 4]]))
   np.testing.assert_allclose(i1, [0, 4])
   np.testing.assert_allclose(i2, [1, 2])
 
@@ -289,11 +519,8 @@ def test_Charge_matmul(chargetype, B0, B1):
   q3 = chargetype(C3)
 
   Q = q1 @ q2 @ q3
-  Q_ = BaseCharge(
-      np.stack([C1, C2, C3], axis=0),
-      charge_labels=None,
-      charge_types=[chargetype] * 3)
-  assert np.all(Q.charges == Q_.charges)
+  Q_ = BaseCharge([C1, C2, C3], charge_types=[[chargetype] for _ in range(3)])
+  np.testing.assert_allclose(np.stack(Q.charges), np.stack(Q_.charges))
 
 
 def test_BaseCharge_matmul_raises():
@@ -323,9 +550,10 @@ def test_Charge_identity(chargetype, B0, B1, identity):
   q3 = chargetype(C3)
 
   Q = q1 @ q2 @ q3
-  eye = Q.identity_charges
-  np.testing.assert_allclose(eye.unique_charges, identity)
+  eye = Q.identity_charges(dim=2)
+  np.testing.assert_allclose(eye.charges[0], identity)
   assert eye.num_symmetries == 3
+  assert len(eye) == 2
 
 
 @pytest.mark.parametrize("n", list(range(2, 20)))
@@ -335,7 +563,7 @@ def test_zncharge_dual_invariant(n):
   charges = np.random.randint(0, n, D).astype(np.int16)
   a = ZNCharge(n)(charges)
   b = a.dual(True)
-  np.testing.assert_allclose((b.charges + a.charges) % n, np.zeros((1, D)))
+  np.testing.assert_allclose((b.charges[0] + a.charges[0]) % n, np.zeros((D)))
 
 
 @pytest.mark.parametrize('chargetype, B0, B1, sign', [(U1Charge, -5, 5, -1),
@@ -349,7 +577,9 @@ def test_Charge_mul(chargetype, B0, B1, sign):
   q2 = chargetype(C2)
   q = q1 @ q2
   res = q * True
-  np.testing.assert_allclose(res.charges, sign * np.stack([C1, C2]))
+  res.expand_charge_types()
+  np.testing.assert_allclose(
+      np.stack(res.charges, axis=0), sign * np.stack([C1, C2], axis=0))
 
 
 @pytest.mark.parametrize('n', list(range(2, 12)))
@@ -363,7 +593,9 @@ def test_Charge_mul_zncharge(n):
   q2 = chargetype(C2)
   q = q1 @ q2
   res = q * True
-  np.testing.assert_allclose(res.charges, (n - np.stack([C1, C2])) % n)
+  res.expand_charge_types()
+  np.testing.assert_allclose(
+      np.stack(res.charges), (n - np.stack([C1, C2])) % n)
 
 
 def test_fuse_charges():
@@ -380,7 +612,7 @@ def test_fuse_charges():
   np_flows[flows] = -1
   fused = fuse_charges(charges, flows)
   np_fused = fuse_ndarrays([c * f for c, f in zip(np_charges, np_flows)])
-  np.testing.assert_allclose(np.squeeze(fused.charges), np_fused)
+  np.testing.assert_allclose(fused.charges[0], np_fused)
 
 
 def test_fuse_charges_raises():
@@ -398,29 +630,42 @@ def test_fuse_charges_raises():
 
 
 def test_reduce():
-  q = np.array([[0, 1, 2, 0, 6, 1, -9, 0, -7], [2, 3, 4, -1, 4, 3, 1, 2, 0]])
+  q = [
+      np.array([0, 1, 2, 0, 6, 1, -9, 0, -7], dtype=np.int16),
+      np.array([2, 3, 4, -1, 4, 3, 1, 2, 0], dtype=np.int16)
+  ]
   Q = BaseCharge(charges=q)
-  target_charge = np.array([[0, 1, 6, -12], [2, 3, 4, 16]])
+  targets = [
+      np.array([0, 1, 6, -12], dtype=np.int16),
+      np.array([2, 3, 4, 16], dtype=np.int16)
+  ]
+  target_charge = BaseCharge(charges=targets)
+
   expected = np.array([[0, 1, 6, 1, 0], [2, 3, 4, 3, 2]])
-  res, locs = Q.reduce(target_charge, return_locations=True)
-  np.testing.assert_allclose(res.charges, expected)
+  res, locs = Q.reduce(
+      target_charge, return_locations=True, return_type='charges')
+  res.expand_charge_types()
+  np.testing.assert_allclose(np.stack(res.charges, axis=0), expected)
   np.testing.assert_allclose(locs, [0, 1, 4, 5, 7])
 
 
 def test_getitem():
-  q1 = np.array([0, 1, 2, 0, 6, 1, -9, 0, -7])
-  q2 = np.array([2, 3, 4, -1, 4, 3, 1, 2, 0])
+  q1 = np.array([0, 1, 2, 0, 6, 1, -9, 0, -7], dtype=np.int16)
+  q2 = np.array([2, 3, 4, -1, 4, 3, 1, 2, 0], dtype=np.int16)
   Q1 = U1Charge(charges=q1)
   Q2 = U1Charge(charges=q2)
   Q = Q1 @ Q2
   t1 = Q[5]
+  t1.expand_charge_types()
   np.testing.assert_allclose(t1.charges, [[1], [3]])
-  assert np.all([t1.charge_types[n] == U1Charge for n in range(2)])
+  assert np.all([t1.charge_types[n][0] is U1Charge for n in range(2)])
   t2 = Q[[2, 5, 7]]
-  assert np.all([t2.charge_types[n] == U1Charge for n in range(2)])
+  t2.expand_charge_types()
+  assert np.all([t2.charge_types[n][0] == U1Charge for n in range(2)])
   np.testing.assert_allclose(t2.charges, [[2, 1, 0], [4, 3, 2]])
   t3 = Q[[5, 2, 7]]
-  assert np.all([t3.charge_types[n] == U1Charge for n in range(2)])
+  t3.expand_charge_types()
+  assert np.all([t3.charge_types[n][0] == U1Charge for n in range(2)])
   np.testing.assert_allclose(t3.charges, [[1, 2, 0], [3, 4, 2]])
 
 
@@ -430,78 +675,29 @@ def test_isin():
   c2 = U1Charge(np.random.randint(-5, 5, 1000, dtype=np.int16))
 
   c = c1 @ c2
-  c3 = np.array([[-1, 0, 1], [-1, 0, 1]])
+  c3 = U1Charge(np.array([-1, 0, 1], dtype=np.int16)) @ U1Charge(
+      np.array([-1, 0, 1], dtype=np.int16))
+  expected = np.array([[-1, 0, 1], [-1, 0, 1]])
+
   n = c.isin(c3)
   for m in np.nonzero(n)[0]:
-    charges = c[m].charges
+    charges = c[m]
+    charges.expand_charge_types()
+    cs = np.stack(charges.charges, axis=0)
     #pylint: disable=unsubscriptable-object
-    assert np.any(
-        [np.array_equal(charges[:, 0], c3[:, k]) for k in range(c3.shape[1])])
-  for m in np.nonzero(np.logical_not(n))[0]:
-    charges = c[m].charges
-    #pylint: disable=unsubscriptable-object
-    assert not np.any(
-        [np.array_equal(charges[:, 0], c3[:, k]) for k in range(c3.shape[1])])
-
-
-def test_isin_2():
-  np.random.seed(10)
-  c1 = U1Charge(np.random.randint(-5, 5, 1000, dtype=np.int16))
-  c2 = U1Charge(np.random.randint(-5, 5, 1000, dtype=np.int16))
-  c = c1 @ c2
-  c3 = U1Charge(np.array([-1, 0, 1])) @ U1Charge(np.array([-1, 0, 1]))
-  n = c.isin(c3)
-  for m in np.nonzero(n)[0]:
-    charges = c[m].charges
     assert np.any([
-        np.array_equal(charges[:, 0], c3.charges[:, k])
-        for k in range(c3.charges.shape[1])
+        np.array_equal(cs[:, 0], expected[:, k])
+        for k in range(expected.shape[1])
     ])
   for m in np.nonzero(np.logical_not(n))[0]:
-    charges = c[m].charges
+    charges = c[m]
+    charges.expand_charge_types()
+    cs = np.stack(charges.charges, axis=0)
+    #pylint: disable=unsubscriptable-object
     assert not np.any([
-        np.array_equal(charges[:, 0], c3.charges[:, k])
-        for k in range(c3.charges.shape[1])
+        np.array_equal(cs[:, 0], expected[:, k])
+        for k in range(expected.shape[1])
     ])
-
-
-def test_isin_raises():
-
-  class FakeCharge(BaseCharge):
-
-    def __init__(self, charges, charge_labels=None, charge_types=None):
-      super().__init__(charges, charge_labels, charge_types=[type(self)])
-
-    @staticmethod
-    def fuse(charge1, charge2) -> np.ndarray:
-      return np.add.outer(charge1, charge2).ravel()
-
-    @staticmethod
-    def dual_charges(charges) -> np.ndarray:
-      return charges * charges.dtype.type(-1)
-
-    @staticmethod
-    def identity_charge() -> np.ndarray:
-      return np.int16(0)
-
-    @classmethod
-    def random(cls, dimension: int, minval: int, maxval: int) -> np.ndarray:
-      charges = np.random.randint(minval, maxval, dimension, dtype=np.int16)
-      return cls(charges=charges)
-
-  np.random.seed(10)
-  c1 = BaseCharge(
-      np.random.randint(-5, 5, (2, 1000), dtype=np.int16),
-      charge_labels=None,
-      charge_types=[FakeCharge, FakeCharge])
-  c2 = U1Charge(np.array([-1, 0, 1])) @ U1Charge(np.array([-1, 0, 1]))
-  with pytest.raises(TypeError):
-    c1.isin(c2)
-  with pytest.raises(ValueError):
-    c1.isin(np.random.randint(-2, 2, (2, 2, 2)))
-
-  with pytest.raises(ValueError):
-    c1.isin(np.random.randint(-2, 2, (3, 2)))
 
 
 def test_eq_1():
@@ -509,11 +705,15 @@ def test_eq_1():
   c1 = U1Charge(np.array([-2, -1, 0, 1, -1, 3, 4, 5], dtype=np.int16))
   c2 = U1Charge(np.array([-1, 0, 1, 2, 0, 4, 5, 6], dtype=np.int16))
   c = c1 @ c2
-  c3 = np.array([[-1], [0]])
+  c3 = U1Charge(np.array([-1], dtype=np.int16)) @ U1Charge(
+      np.array([0], dtype=np.int16))
+  exp = np.array([[-1], [0]])
   inds = np.nonzero(c == c3)[0]
   np.testing.assert_allclose(inds, [1, 4])
   for i in inds:
-    np.array_equal(c[i].charges, c3)
+    tmp = c[i]
+    tmp.expand_charge_types()
+    np.array_equal(np.stack(tmp.charges), exp)
 
 
 def test_eq_2():
@@ -521,67 +721,33 @@ def test_eq_2():
   c1 = U1Charge(np.array([-2, -1, 0, 1, -1, 3, 4, 5, 1], dtype=np.int16))
   c2 = U1Charge(np.array([-1, 0, 1, 2, 0, 4, 5, 6, 2], dtype=np.int16))
   c = c1 @ c2
-  c3 = np.array([[-1, 1], [0, 2]])
+  c3 = U1Charge(np.array([-1, 1], dtype=np.int16)) @ U1Charge(
+      np.array([0, 2], dtype=np.int16))
+  exp = np.array([[-1, 1], [0, 2]])
   inds = np.nonzero(c == c3)
 
   np.testing.assert_allclose(inds[0][inds[1] == 0], [1, 4])
   np.testing.assert_allclose(inds[0][inds[1] == 1], [3, 8])
   for i, j in zip(inds[0], inds[1]):
-    np.array_equal(c[i].charges, c3[:, j])
-
-
-def test_eq__raises():
-  np.random.seed(10)
-  num_charges = 2
-  charge = BaseCharge(
-      np.random.randint(-2, 3, (num_charges, 30)),
-      charge_types=[U1Charge] * num_charges)
-  with pytest.raises(ValueError):
-    _ = charge == np.random.randint(-1, 1, (num_charges + 1, 2), dtype=np.int16)
-
-
-def test_iter():
-  np.random.seed(10)
-  arr1 = np.array([-2, -1, 0, 1, -1, 3, 4, 5, 1], dtype=np.int16)
-  arr2 = np.array([-1, 0, 1, 2, 0, 4, 5, 6, 2], dtype=np.int16)
-  c1 = U1Charge(arr1)
-  c2 = U1Charge(arr2)
-  c = c1 @ c2
-  m = 0
-  for n in c:
-    np.testing.assert_allclose(n, np.array([arr1[m], arr2[m]]))
-    m += 1
+    tmp = c[i]
+    tmp.expand_charge_types()
+    np.array_equal(np.stack(tmp.charges), exp[:, j])
 
 
 def test_empty():
   num_charges = 4
-  charges = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, 0)),
-      charge_types=[U1Charge] * num_charges)
+  charges = BaseCharge([
+      np.random.randint(-5, 6, 0).astype(np.int16) for _ in range(num_charges)
+  ])
   assert len(charges) == 0
 
 
 def test_init_raises():
-  num_charges = 4
+  charge_types = [[U1Charge, Z2Charge], [U1Charge]]
   with pytest.raises(ValueError):
     BaseCharge(
-        np.random.randint(-5, 6, (num_charges, 10)),
-        charge_types=[U1Charge] * (num_charges - 1))
-
-
-def test_eq_raises():
-  num_charges = 4
-  c1 = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, 10)),
-      charge_types=[U1Charge] * num_charges)
-  c2 = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, 0)),
-      charge_types=[U1Charge] * num_charges)
-  npc = np.empty((num_charges, 0), dtype=np.int16)
-  with pytest.raises(ValueError):
-    c1 == c2
-  with pytest.raises(ValueError):
-    c1 == npc
+        [np.random.randint(-5, 6, 10).astype(np.int16) for _ in range(2)],
+        charge_types=charge_types)
 
 
 def test_zncharge_raises():
