+++ tensornetwork/block_sparse/linalg_test.py
@@ -1,13 +1,13 @@
 import numpy as np
 import pytest
 # pylint: disable=line-too-long
-from tensornetwork.block_sparse.charge import U1Charge, fuse_charges, charge_equal, BaseCharge
-from tensornetwork.block_sparse.index import Index
-from tensornetwork.block_sparse.blocksparsetensor import ChargeArray, BlockSparseTensor
-from tensornetwork.block_sparse.utils import _find_diagonal_sparse_blocks
-from tensornetwork import ncon
+from tensornetwork.block_sparse.charge_new import U1Charge, fuse_charges, charge_equal, BaseCharge
+from tensornetwork.block_sparse.index_new import Index
+from tensornetwork.block_sparse.blocksparsetensor_new import ChargeArray, BlockSparseTensor
+from tensornetwork.block_sparse.utils_new import _find_diagonal_sparse_blocks
+from tensornetwork.ncon_interface import ncon
 # pylint: disable=line-too-long
-from tensornetwork.block_sparse.linalg import norm, diag, reshape, transpose, conj, svd, qr, eigh, eig, inv, sqrt, trace, eye, pinv, zeros, ones, randn, random
+from tensornetwork.block_sparse.linalg_new import norm, diag, reshape, transpose, conj, svd, qr, eigh, eig, inv, sqrt, trace, eye, pinv, zeros, ones, randn, random
 
 np_dtypes = [np.float64, np.complex128]
 np_tensordot_dtypes = [np.float64, np.complex128]
@@ -38,25 +38,29 @@ def test_get_diag(dtype, num_charges, Ds, flow):
   indices = [
       Index(
           BaseCharge(
-              np.random.randint(-2, 3, (num_charges, Ds[n])),
-              charge_types=[U1Charge] * num_charges), flow) for n in range(2)
+              [np.random.randint(-2, 3, Ds[n], dtype=np.int16) for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), flow) for n in range(2)
   ]
   arr = BlockSparseTensor.random(indices, dtype=dtype)
   fused = fuse_charges(arr.flat_charges, arr.flat_flows)
-  inds = np.nonzero(fused == np.zeros((num_charges, 1), dtype=np.int16))[0]
+  inds = np.nonzero(fused == fused.identity_charges(1))[0]
   # pylint: disable=no-member
   left, _ = np.divmod(inds, Ds[1])
+  [i._charges[0].expand_charge_types() for i in indices]
   unique = np.unique(
-      np_flow * (indices[0]._charges[0].charges[:, left]), axis=1)
+      np_flow * (np.stack(indices[0]._charges[0].charges)[:, left]), axis=1)
   diagonal = diag(arr)
+  
   sparse_blocks, _, block_shapes = _find_diagonal_sparse_blocks(
       arr.flat_charges, arr.flat_flows, 1)
   data = np.concatenate([
       np.diag(np.reshape(arr.data[sparse_blocks[n]], block_shapes[:, n]))
       for n in range(len(sparse_blocks))
   ])
+  actual = diagonal.flat_charges[0].unique()
+  actual.expand_charge_types()
   np.testing.assert_allclose(data, diagonal.data)
-  np.testing.assert_allclose(unique, diagonal.flat_charges[0].unique_charges)
+  np.testing.assert_allclose(unique, np.stack(actual.charges))
 
 
 @pytest.mark.parametrize('dtype', np_dtypes)
@@ -67,8 +71,8 @@ def test_get_empty_diag(dtype, num_charges, Ds):
   indices = [
       Index(
           BaseCharge(
-              np.random.randint(-2, 3, (num_charges, Ds[n])),
-              charge_types=[U1Charge] * num_charges), False) for n in range(2)
+              [np.random.randint(-2, 3, Ds[n], dtype=np.int16) for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), False) for n in range(2)
   ]
   arr = BlockSparseTensor.random(indices, dtype=dtype)
   diagonal = diag(arr)
@@ -84,9 +88,8 @@ def test_create_diag(dtype, num_charges, flow):
   np.random.seed(10)
   D = 200
   index = Index(
-      BaseCharge(
-          np.random.randint(-2, 3, (num_charges, D)),
-          charge_types=[U1Charge] * num_charges), flow)
+      BaseCharge([np.random.randint(-2, 3, D, dtype=np.int16) for _ in range(num_charges)],
+                 charge_types=[[U1Charge]] * num_charges), flow)
 
   arr = ChargeArray.random([index], dtype=dtype)
   diagarr = diag(arr)
@@ -111,9 +114,8 @@ def test_diag_raises():
   rank = len(Ds)
   indices = [
       Index(
-          BaseCharge(
-              np.random.randint(-2, 3, (1, Ds[n])), charge_types=[U1Charge]),
-          False) for n in range(rank)
+          BaseCharge([np.random.randint(-2, 3, Ds[n], dtype=np.int16)],
+                     charge_types=[[U1Charge]]), False) for n in range(rank)
   ]
   arr = BlockSparseTensor.random(indices)
   chargearr = ChargeArray.random([indices[0], indices[1]])
@@ -213,9 +215,8 @@ def test_svd_prod(dtype, Ds, R1, num_charges):
   np.random.seed(10)
   R = len(Ds)
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, Ds[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([np.random.randint(-5, 6, Ds[n], dtype=np.int16) for _ in range(num_charges)],
+                 charge_types=[[U1Charge]] * num_charges) for n in range(R)
   ]
   flows = [True] * R
   A = BlockSparseTensor.random([Index(charges[n], flows[n]) for n in range(R)],
@@ -241,9 +242,8 @@ def test_svd_singvals(dtype, Ds, R1, num_charges):
   np.random.seed(10)
   R = len(Ds)
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, Ds[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([np.random.randint(-5, 6, Ds[n], dtype=np.int16) for _ in range(num_charges)],
+                 charge_types=[[U1Charge]] * num_charges) for n in range(R)
   ]
   flows = [True] * R
   A = BlockSparseTensor.random([Index(charges[n], flows[n]) for n in range(R)],
@@ -270,9 +270,8 @@ def test_qr_prod(dtype, Ds, R1, mode, num_charges):
   np.random.seed(10)
   R = len(Ds)
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, Ds[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([np.random.randint(-5, 6, Ds[n], dtype=np.int16) for _ in range(num_charges)],
+                 charge_types=[[U1Charge]] * num_charges) for n in range(R)
   ]
   flows = [True] * R
   A = BlockSparseTensor.random([Index(charges[n], flows[n]) for n in range(R)],
@@ -296,9 +295,8 @@ def test_qr_raises():
   R1 = 1
   R = len(Ds)
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, Ds[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([np.random.randint(-5, 6, Ds[n], dtype=np.int16) for _ in range(num_charges)],
+                 charge_types=[[U1Charge]] * num_charges) for n in range(R)
   ]
   flows = [True] * R
   A = BlockSparseTensor.random([Index(charges[n], flows[n]) for n in range(R)],
@@ -317,9 +315,12 @@ def test_eigh_prod(dtype, Ds, num_charges):
   np.random.seed(10)
   R = len(Ds)
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, Ds[n]), dtype=np.int16),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([
+          np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+          for _ in range(num_charges)
+      ],
+                 charge_types=[[U1Charge]] * num_charges)
+      for n in range(R)
   ]
   flows = [False] * R
   inds = [Index(charges[n], flows[n]) for n in range(R)]
@@ -341,9 +342,12 @@ def test_eigh_raises():
   D = 20
   R = 3
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([
+          np.random.randint(-5, 6, D, dtype=np.int16)
+          for _ in range(num_charges)
+      ],
+                 charge_types=[[U1Charge]] * num_charges)
+      for n in range(R)
   ]
   flows = [False] * R
   inds = [Index(charges[n], flows[n]) for n in range(R)]
@@ -359,8 +363,8 @@ def test_inv(dtype, num_charges):
   R = 2
   D = 10
   charge = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   flows = [True, False]
   A = BlockSparseTensor.random([Index(charge, flows[n]) for n in range(R)],
                                (-0.5, 0.5),
@@ -388,8 +392,8 @@ def test_inv_raises():
   R = 3
   D = 10
   charge = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   A = BlockSparseTensor.random([Index(charge, False) for n in range(R)],
                                (-0.5, 0.5))
   with pytest.raises(ValueError):
@@ -403,9 +407,12 @@ def test_eig_prod(dtype, Ds, num_charges):
   np.random.seed(10)
   R = len(Ds)
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, Ds[n]), dtype=np.int16),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([
+          np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+          for _ in range(num_charges)
+      ],
+                 charge_types=[[U1Charge]] * num_charges)
+      for n in range(R)
   ]
   flows = [False] * R
   inds = [Index(charges[n], flows[n]) for n in range(R)]
@@ -425,9 +432,12 @@ def test_eig_raises():
   D = 20
   R = 3
   charges = [
-      BaseCharge(
-          np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-          charge_types=[U1Charge] * num_charges) for n in range(R)
+      BaseCharge([
+          np.random.randint(-5, 6, D, dtype=np.int16)
+          for _ in range(num_charges)
+      ],
+                 charge_types=[[U1Charge]] * num_charges)
+      for n in range(R)
   ]
   flows = [False] * R
   inds = [Index(charges[n], flows[n]) for n in range(R)]
@@ -446,9 +456,11 @@ def test_sqrt(dtype, num_charges, Ds):
   flows = np.random.choice([True, False], replace=True, size=R)
   indices = [
       Index(
-          BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds[n]), dtype=np.int16),
-              charge_types=[U1Charge] * num_charges), flows[n])
+          BaseCharge([
+              np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+              for _ in range(num_charges)
+          ],
+                     charge_types=[[U1Charge]] * num_charges), flows[n])
       for n in range(R)
   ]
   arr = BlockSparseTensor.random(indices, dtype=dtype)
@@ -461,8 +473,8 @@ def test_sqrt(dtype, num_charges, Ds):
 @pytest.mark.parametrize('D', [0, 10])
 def test_eye(dtype, num_charges, D):
   charge = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   flow = False
   index = Index(charge, flow)
   A = eye(index, dtype=dtype)
@@ -480,8 +492,8 @@ def test_trace_matrix(dtype, num_charges, D):
   np.random.seed(10)
   R = 2
   charge = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   flows = [True, False]
   matrix = BlockSparseTensor.random([Index(charge, flows[n]) for n in range(R)],
                                     dtype=dtype)
@@ -495,12 +507,14 @@ def test_trace_matrix(dtype, num_charges, D):
 @pytest.mark.parametrize('D1, D2', [(10, 12), (0, 10)])
 def test_trace_tensor(dtype, num_charges, D1, D2):
   np.random.seed(10)
-  charge1 = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D1), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
-  charge2 = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D2), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+  charge1 = BaseCharge([
+      np.random.randint(-5, 6, D1, dtype=np.int16) for _ in range(num_charges)
+  ],
+                       charge_types=[[U1Charge]] * num_charges)
+  charge2 = BaseCharge([
+      np.random.randint(-5, 6, D2, dtype=np.int16) for _ in range(num_charges)
+  ],
+                       charge_types=[[U1Charge]] * num_charges)
   indices = [Index(charge1, False), Index(charge2, False), Index(charge1, True)]
   tensor = BlockSparseTensor.random(indices, dtype=dtype)
   res = trace(tensor, (0, 2))
@@ -514,15 +528,15 @@ def test_trace_raises(num_charges):
   np.random.seed(10)
   D = 20
   charge1 = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   A1 = BlockSparseTensor.random([Index(charge1, False)])
   with pytest.raises(ValueError):
     trace(A1)
 
   charge2 = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D + 1), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+    [np.random.randint(-5, 6, D + 1, dtype=np.int16) for _ in range(num_charges)],
+    charge_types=[[U1Charge]] * num_charges)
   indices = [
       Index(charge1, False),
       Index(charge2, False),
@@ -544,8 +558,8 @@ def test_pinv(dtype, num_charges):
   R = 2
   D = 10
   charge = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   flows = [True, False]
   A = BlockSparseTensor.random([Index(charge, flows[n]) for n in range(R)],
                                (-0.5, 0.5),
@@ -573,8 +587,8 @@ def test_pinv_raises():
   R = 3
   D = 10
   charge = BaseCharge(
-      np.random.randint(-5, 6, (num_charges, D), dtype=np.int16),
-      charge_types=[U1Charge] * num_charges)
+      [np.random.randint(-5, 6, D, dtype=np.int16) for _ in range(num_charges)],
+      charge_types=[[U1Charge]] * num_charges)
   A = BlockSparseTensor.random([Index(charge, False) for n in range(R)],
                                (-0.5, 0.5))
   with pytest.raises(ValueError):
@@ -591,8 +605,9 @@ def test_tn_zeros(dtype, num_charges):
   indices = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds[n])),
-              charge_types=[U1Charge] * num_charges), flows[n])
+              [np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+               for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), flows[n])
       for n in range(rank)
   ]
   arr = zeros(indices, dtype=dtype)
@@ -613,8 +628,9 @@ def test_tn_ones(dtype, num_charges):
   indices = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds[n])),
-              charge_types=[U1Charge] * num_charges), flows[n])
+              [np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+               for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), flows[n])
       for n in range(rank)
   ]
 
@@ -636,8 +652,9 @@ def test_tn_random(dtype, num_charges):
   indices = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds[n])),
-              charge_types=[U1Charge] * num_charges), flows[n])
+              [np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+               for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), flows[n])
       for n in range(rank)
   ]
   arr = random(indices, dtype=dtype)
@@ -658,8 +675,9 @@ def test_tn_randn(dtype, num_charges):
   indices = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds[n])),
-              charge_types=[U1Charge] * num_charges), flows[n])
+              [np.random.randint(-5, 6, Ds[n], dtype=np.int16)
+               for _ in range(num_charges)],
+              charge_types=[[U1Charge]] * num_charges), flows[n])
       for n in range(rank)
   ]
   arr = randn(indices, dtype=dtype)
