+++ tensornetwork/block_sparse/tensordot_test.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pytest
 # pylint: disable=line-too-long
-from tensornetwork.block_sparse.charge import U1Charge, charge_equal, BaseCharge
-from tensornetwork.block_sparse.index import Index
-from tensornetwork.block_sparse.blocksparsetensor import BlockSparseTensor, tensordot, outerproduct
+from tensornetwork.block_sparse.charge_new import U1Charge, charge_equal, BaseCharge
+from tensornetwork.block_sparse.index_new import Index
+from tensornetwork.block_sparse.blocksparsetensor_new import BlockSparseTensor, tensordot, outerproduct
 from tensornetwork import ncon
 
 np_dtypes = [np.float64, np.complex128]
@@ -16,18 +16,18 @@ def get_contractable_tensors(R1, R2, cont, dtype, num_charges, DsA, Dscomm,
   assert R2 >= cont
   chargesA = [
       BaseCharge(
-          np.random.randint(-5, 5, (num_charges, DsA[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(R1 - cont)
+        [np.random.randint(-5, 5, DsA[n], dtype=np.int16) for _ in range(num_charges)],
+        charge_types=[[U1Charge]] * num_charges) for n in range(R1 - cont)
   ]
   commoncharges = [
       BaseCharge(
-          np.random.randint(-5, 5, (num_charges, Dscomm[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(cont)
+        [np.random.randint(-5, 5, Dscomm[n], dtype=np.int16)for _ in range(num_charges)],
+        charge_types=[[U1Charge]] * num_charges) for n in range(cont)
   ]
   chargesB = [
       BaseCharge(
-          np.random.randint(-5, 5, (num_charges, DsB[n])),
-          charge_types=[U1Charge] * num_charges) for n in range(R2 - cont)
+        [np.random.randint(-5, 5, DsB[n], dtype=np.int16) for _ in range(num_charges)],
+        charge_types=[[U1Charge]] * num_charges) for n in range(R2 - cont)
   ]
   #contracted indices
   indsA = np.random.choice(np.arange(R1), cont, replace=False)
@@ -69,15 +69,15 @@ def test_outerproduct(dtype, num_legs, num_charges):
   is1 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds1[n])),
-              charge_types=[U1Charge] * num_charges), False)
+            [np.random.randint(-5, 6, Ds1[n], dtype=np.int16) for _ in range(num_charges)],
+            charge_types=[[U1Charge]] * num_charges), False)
       for n in range(num_legs)
   ]
   is2 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds2[n])),
-              charge_types=[U1Charge] * num_charges), False)
+            [np.random.randint(-5, 6, Ds2[n], dtype=np.int16) for _ in range(num_charges)],
+            charge_types=[[U1Charge]] * num_charges), False)
       for n in range(num_legs)
   ]
   a = BlockSparseTensor.random(is1, dtype=dtype)
@@ -100,15 +100,15 @@ def test_outerproduct_transpose(dtype, num_legs, num_charges):
   is1 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds1[n])),
-              charge_types=[U1Charge] * num_charges), False)
+            [np.random.randint(-5, 6, Ds1[n], dtype=np.int16) for _ in range(num_charges)],
+            charge_types=[[U1Charge]] * num_charges), False)
       for n in range(num_legs)
   ]
   is2 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds2[n])),
-              charge_types=[U1Charge] * num_charges), False)
+            [np.random.randint(-5, 6, Ds2[n], dtype=np.int16) for _ in range(num_charges)],
+            charge_types=[[U1Charge]] * num_charges), False)
       for n in range(num_legs)
   ]
   o1 = np.arange(num_legs)
@@ -136,15 +136,15 @@ def test_outerproduct_transpose_reshape(dtype, num_legs, num_charges):
   is1 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds1[n])),
-              charge_types=[U1Charge] * num_charges), False)
+            [np.random.randint(-5, 6, Ds1[n], dtype=np.int16) for _ in range(num_charges)],
+            charge_types=[[U1Charge]] * num_charges), False)
       for n in range(num_legs)
   ]
   is2 = [
       Index(
           BaseCharge(
-              np.random.randint(-5, 6, (num_charges, Ds2[n])),
-              charge_types=[U1Charge] * num_charges), False)
+            [np.random.randint(-5, 6, Ds2[n], dtype=np.int16) for _ in range(num_charges)],
+            charge_types=[[U1Charge]] * num_charges), False)
       for n in range(num_legs)
   ]
   o1 = np.arange(num_legs)
@@ -254,12 +254,12 @@ def test_tensordot_reshape(dtype, num_charges):
   R1 = 4
   R2 = 4
 
-  q = np.random.randint(-5, 5, (num_charges, 10), dtype=np.int16)
+  q = [np.random.randint(-5, 5, 10, dtype=np.int16) for _ in range(num_charges)]
   charges1 = [
-      BaseCharge(q, charge_types=[U1Charge] * num_charges) for n in range(R1)
+      BaseCharge(q, charge_types=[[U1Charge]] * num_charges) for n in range(R1)
   ]
   charges2 = [
-      BaseCharge(q, charge_types=[U1Charge] * num_charges) for n in range(R2)
+      BaseCharge(q, charge_types=[[U1Charge]] * num_charges) for n in range(R2)
   ]
   flowsA = np.asarray([False] * R1)
   flowsB = np.asarray([True] * R2)
